backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
pydantic>=2.6.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
selenium==4.15.2
webdriver-manager==4.0.1
openai==1.3.7
python-dotenv==1.0.0
sqlalchemy==2.0.43
alembic==1.12.1
google-generativeai>=0.4.0
